csv upload moved to campaign edit screen. now creates the entry in VotersList only after the list has been uploaded and saved.
git-svn-id: https://impact-dialing.svn.beanstalkapp.com/webapp/trunk@369 2f5aa9f8-122e-46c4-b90d-7e846bb7426a

+++ app/controllers/voter_lists_controller.rb
@@ -9,37 +9,21 @@ class VoterListsController < ClientController
     end
 
     uploaded_file = params[:upload]["datafile"]
-    separator     = params[:seperator]=="tab" ? "\t" : ","
 
-    @voter_list             = VoterList.new
-    @voter_list.campaign_id = params[:campaign_id]
-    @voter_list.name        = params[:list_name]
-    @voter_list.user_id     = session[:user]
-    unless @voter_list.valid?
-      flash[:error] = @voter_list.errors.full_messages
-      return
-    end
-    @voter_list.save!
-
-    saved_filename = write_csv_file(uploaded_file, @voter_list)
-    save_csv_filename_to_session(saved_filename, separator, @voter_list)
+    @separator         = separator_from_file_extension(uploaded_file.original_filename)
+    saved_filename = write_csv_file(uploaded_file)
+    save_csv_filename_to_session(saved_filename)
 
     @system_column_headers = VoterList::VOTER_DATA_COLUMNS.zip(VoterList::VOTER_DATA_COLUMNS)
     @system_column_headers = [["Not available", nil]].concat @system_column_headers
-    @csv_column_headers    = FasterCSV.parse(uploaded_file.readline, :col_sep => separator).first
+    @csv_column_headers    = FasterCSV.parse(uploaded_file.readline, :col_sep => @separator).first
 
     render "column_mapping"
   end
 
   def add_to_db
-    id         = params["id"].to_i
-    voter_list = VoterList.find(id)
-    unless voter_list.user_id.to_s == session[:user].to_s
-      flash[:error] = "You are not authorized to edit this list"
-      redirect_to new_campaign_voter_list_path(@campaign.id)
-      return
-    end
 
+    separator = params["separator"]
     csv_to_system_map = params["csv_to_system_map"]
     phone_column      = csv_to_system_map.values.map(&:upcase).index("PHONE")
     unless phone_column.present?
@@ -48,15 +32,27 @@ class VoterListsController < ClientController
       return
     end
 
-    csv_filename      = session[:voters_list_uploads][id]["filename"]
-    separator         = session[:voters_list_uploads][id]["separator"]
+    csv_filename      = session[:voters_list_upload]["filename"]
     uploaded_filename = temp_file_path(csv_filename)
 
-    @result = voter_list.import_leads(csv_to_system_map,
+    @voter_list             = VoterList.new
+    @voter_list.name        = params[:voter_list_name]
+    @voter_list.campaign_id = params[:campaign_id]
+    @voter_list.user_id     = session[:user]
+    @voter_list.state       = VoterList::States::INITIAL
+    unless @voter_list.valid?
+      flash[:error] = @voter_list.errors.full_messages
+      return
+    end
+    @voter_list.save!
+
+    @result = @voter_list.import_leads(csv_to_system_map,
                                       uploaded_filename,
                                       separator)
+    
+    @voter_list.update_attribute(:state, VoterList::States::VALID)
     File.unlink uploaded_filename
-    session[:voters_list_uploads].delete id
+    session[:voters_list_upload] = nil
     render :show
   end
 
@@ -72,9 +68,9 @@ class VoterListsController < ClientController
     @campaign = Campaign.find(params[:campaign_id])
   end
 
-  def write_csv_file(uploaded_file, voter_list)
+  def write_csv_file(uploaded_file)
     uploaded_file = params[:upload]["datafile"]
-    csv_filename  = "uploaded_list_#{voter_list.id}_#{rand(100)}"
+    csv_filename  = "#{uploaded_file.original_filename}_#{Time.now.to_i}_#{rand(999)}"
     File.open(temp_file_path(csv_filename), "w") do |f|
       f.write(uploaded_file.read)
       f.flush
@@ -83,15 +79,18 @@ class VoterListsController < ClientController
     csv_filename
   end
 
-  def save_csv_filename_to_session(csv_filename, separator, voter_list)
-    session[:voters_list_uploads]                ||= {}
-    session[:voters_list_uploads][voter_list.id] = {
+  def save_csv_filename_to_session(csv_filename)
+    session[:voters_list_upload] = {
         "filename"    => csv_filename,
-        "separator"   => separator,
         "upload_time" => Time.now}
   end
 
   def temp_file_path(filename)
     "#{Rails.root}/tmp/#{filename}"
+  end
+
+  private
+  def separator_from_file_extension(filename)
+    (File.extname(filename).downcase.include?('.csv')) ? ',' : '\t'
   end
 end
